decompose manager interfaces
