+++ pacco/manager/remote_manager.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Optional
 
 import yaml
 
-from pacco.manager.classes_interface import PackageManager
+from pacco.manager.package_manager import PackageManager
 from pacco.manager.classes_file_based import PackageManagerFileBased
 from pacco.client.clients import LocalClient, NexusFileClient, FileBasedClientAbstract
